// src/types.ts
export interface IDimensions {
    initialWidth?: number
    initialHeight?: number
    dpr?: number

    canvasWidth?: number
    canvasHeight?: number
    
    scale?: number
    width?: number
    height?: number

    /**
     * Used to fix inaccuracies at the high page zoom
     */
    scaleOffset?: number
}

export type IPosition = { x: number, y: number }